+++ ts/src/error.ts
@@ -58,7 +58,7 @@ const LangErrorCode = {
 
   // Constraints.
   ConstraintMut: 140,
-  ConstraintBelongsTo: 141,
+  ConstraintHasOne: 141,
   ConstraintSigner: 142,
   ConstraintRaw: 143,
   ConstraintOwner: 144,
@@ -119,7 +119,7 @@ const LangErrorMessage = new Map([
 
   // Constraints.
   [LangErrorCode.ConstraintMut, "A mut constraint was violated"],
-  [LangErrorCode.ConstraintBelongsTo, "A belongs_to constraint was violated"],
+  [LangErrorCode.ConstraintHasOne, "A has_one constraint was violated"],
   [LangErrorCode.ConstraintSigner, "A signer constraint was violated"],
   [LangErrorCode.ConstraintRaw, "A raw constraint as violated"],
   [LangErrorCode.ConstraintOwner, "An owner constraint was violated"],